Refactor: Move all service implementations to two core services
Prepared to make Search-AzureResourceGraph public without exporting it. Considering moving it to its own module.
- Removed the downloadDependencies script as its functionality has been integrated into the build process.
