+++ Docs/Help/Get-BicepConfig.md
@@ -33,31 +33,34 @@ Get-BicepConfig -Path <String> [-AsString] [<CommonParameters>]
 ```
 
 ## DESCRIPTION
-Command to get the bicep configuration in use for a specific Bicep file.
-Will return path to the bicepconfig.json file as well as the current settings.
+Command to get the bicep configuration in use for a specific Bicep file. Will return path to the bicepconfig.json file as well as the current settings.
 
 ## EXAMPLES
 
-### Example 1 - Get bicep configuration for a bicep file
+### Example 1
+Get bicep configuration for a bicep file
 ```
 Get-BicepConfig -Path .\storage.bicep
 ```
 
-### Example 2 - Get the merged bicep configuration for a bicep file
+### Example 2
+Get the merged bicep configuration for a bicep file
 ```
 Get-BicepConfig -Path .\storage.bicep -Merged
 ```
 
 Returns the path to the bicepconfig.json file in use, and the merged settings (default + local file).
 
-### Example 3 - Get the local bicep configuration for a bicep file
+### Example 3
+Get the local bicep configuration for a bicep file
 ```
 Get-BicepConfig -Path .\storage.bicep -Merged
 ```
 
 Returns the path to the bicepconfig.json file in use, and the settings in the local bicepconfig.json.
 
-### Example 4 - Get the default bicep configuration
+### Example 4
+Get the default bicep configuration
 ```
 Get-BicepConfig -Default
 ```
@@ -65,6 +68,21 @@ Get-BicepConfig -Default
 Returns the default settings.
 
 ## PARAMETERS
+
+### -AsString
+Output the configuration as a string instead of an object.
+
+```yaml
+Type: SwitchParameter
+Parameter Sets: (All)
+Aliases:
+
+Required: False
+Position: Named
+Default value: False
+Accept pipeline input: False
+Accept wildcard characters: False
+```
 
 ### -Default
 Returns the default bicep configuration
@@ -122,21 +140,6 @@ Aliases:
 Required: True
 Position: Named
 Default value: None
-Accept pipeline input: False
-Accept wildcard characters: False
-```
-
-### -AsString
-Output result as string.
-
-```yaml
-Type: SwitchParameter
-Parameter Sets: (All)
-Aliases:
-
-Required: False
-Position: Named
-Default value: False
 Accept pipeline input: False
 Accept wildcard characters: False
 ```
