+++ PSBicep.Core/BicepHelper.cs
@@ -1,4 +1,5 @@
 using System;
+using System.Collections.Generic;
 using System.Linq;
 using Azure.Deployments.Core.Comparers;
 using Bicep.Core.Resources;
@@ -9,12 +10,11 @@ namespace PSBicep.Core;
 
 internal static class BicepHelper
 {
-    internal static ResourceTypeReference ResolveBicepTypeDefinition(string fullyQualifiedType, AzResourceTypeLoader azResourceTypeLoader, ILogger? logger = null)
+    internal static ResourceTypeReference ResolveBicepTypeDefinition(string fullyQualifiedType, AzResourceTypeLoader azResourceTypeLoader, ILogger? logger = null, int skip = 0, bool avoidPreview = false)
     {
-        var matchedType = azResourceTypeLoader.GetAvailableTypes()
-            .Where(x => StringComparer.OrdinalIgnoreCase.Equals(fullyQualifiedType, x.FormatType()))
-            .OrderByDescending(x => x.ApiVersion ?? "", ApiVersionComparer.Instance)
+        var matchedType = GetApiVersions(fullyQualifiedType, azResourceTypeLoader, logger, skip, avoidPreview)
             .FirstOrDefault();
+
         if (matchedType is null || matchedType.ApiVersion is null)
         {
             var message = $"Failed to find a Bicep type definition for resource of type \"{fullyQualifiedType}\".";
@@ -25,7 +25,42 @@ internal static class BicepHelper
         return matchedType;
     }
 
-    internal static string[] GetApiVersions(ResourceTypeReference typeReference, AzResourceTypeLoader azResourceTypeLoader)
+    internal static IEnumerable<ResourceTypeReference> GetApiVersions(string fullyQualifiedType, AzResourceTypeLoader azResourceTypeLoader, ILogger? logger = null, int skip = 0, bool avoidPreview = false)
+    {
+        var allMatchedTypes = azResourceTypeLoader.GetAvailableTypes()
+            .Where(x => StringComparer.OrdinalIgnoreCase.Equals(fullyQualifiedType, x.FormatType()))
+            .Where(x => x.ApiVersion is not null);
+
+        if (avoidPreview)
+        {
+            var NonPreviewMatchedTypes = allMatchedTypes
+                .Where(x => !x.ApiVersion!.EndsWith("-preview", StringComparison.OrdinalIgnoreCase));
+
+            if (NonPreviewMatchedTypes.Any())
+            {
+                allMatchedTypes = NonPreviewMatchedTypes;
+            }
+        }
+
+        if (skip > 0 && allMatchedTypes.Count() > skip)
+        {
+            allMatchedTypes = allMatchedTypes.Take(skip);
+        }
+
+        var orderedType = allMatchedTypes
+            .OrderByDescending(x => x.ApiVersion ?? "", ApiVersionComparer.Instance);
+
+        if (orderedType is null)
+        {
+            var message = $"Failed to find a Bicep type definition for resource of type \"{fullyQualifiedType}\".";
+            logger?.LogCritical("{message}", message);
+            throw new InvalidOperationException(message);
+        }
+
+        return orderedType;
+    }
+
+    internal static string[] GetApiVersions(ResourceTypeReference typeReference, AzResourceTypeLoader azResourceTypeLoader, int skip = 0, bool avoidPreview = false)
     {
         return azResourceTypeLoader.GetAvailableTypes()
             .Where(x => StringComparer.OrdinalIgnoreCase.Equals(typeReference.FormatType(), x.FormatType()))
