+++ PSBicep/Commands/BaseCommand.cs
@@ -1,21 +1,20 @@
 ﻿using System.Management.Automation;
-using PSBicep.Core;
 
 namespace PSBicep.Commands;
 
 public class BaseCommand : PSCmdlet
 {
     protected string name;
-    protected BicepService bicepService;
+    protected Core.PSBicep psBicep;
 
     protected override void BeginProcessing()
     {
         base.BeginProcessing();
-        bicepService = new BicepService(this);
+        psBicep = new Core.PSBicep(this);
     }
 
     protected void SetAuthentication(string token)
     {
-        bicepService.authentication.SetAuthentication(token);
+        psBicep.registryService.SetAuthentication(token);
     }
 }
