+++ build.ps1
@@ -1,6 +1,6 @@
 <#
     .DESCRIPTION
-        Bootstrap and build script for PowerShell module CI/CD pipeline
+        Bootstrap and build script for PowerShell module CI/CD pipeline.
 
     .PARAMETER Tasks
         The task or tasks to run. The default value is '.' (runs the default task).
@@ -56,6 +56,19 @@
 
     .PARAMETER AutoRestore
         Not yet written.
+
+    .PARAMETER UseModuleFast
+        Specifies to use ModuleFast instead of PowerShellGet to resolve dependencies
+        faster.
+
+    .PARAMETER UsePSResourceGet
+        Specifies to use PSResourceGet instead of PowerShellGet to resolve dependencies
+        faster. This can also be configured in Resolve-Dependency.psd1.
+
+    .PARAMETER UsePowerShellGetCompatibilityModule
+        Specifies to use the compatibility module PowerShellGet. This parameter
+        only works then the method of downloading dependencies is PSResourceGet.
+        This can also be configured in Resolve-Dependency.psd1.
 #>
 [CmdletBinding()]
 param
@@ -121,7 +134,19 @@ param
 
     [Parameter()]
     [System.Management.Automation.SwitchParameter]
-    $AutoRestore
+    $AutoRestore,
+
+    [Parameter()]
+    [System.Management.Automation.SwitchParameter]
+    $UseModuleFast,
+
+    [Parameter()]
+    [System.Management.Automation.SwitchParameter]
+    $UsePSResourceGet,
+
+    [Parameter()]
+    [System.Management.Automation.SwitchParameter]
+    $UsePowerShellGetCompatibilityModule
 )
 
 <#
@@ -142,6 +167,15 @@ process
 
     # Execute the Build process from the .build.ps1 path.
     Push-Location -Path $PSScriptRoot -StackName 'BeforeBuild'
+
+    if (Test-Path -Path 'Sampler')
+    {
+        # We are in the Sampler project, load functions instead of Sampler module.
+        Get-ChildItem -Path "Sampler/P*/*.ps1" |
+            ForEach-Object -Process {
+                . $_.FullName
+            }
+    }
 
     try
     {
@@ -179,7 +213,7 @@ process
                             ConvertFrom-Yaml -Yaml (Get-Content -Raw $configFile)
                         }
 
-                        # Native Support for JSON and JSONC (by Removing comments)
+                        # Support for JSON and JSONC (by Removing comments) when module PowerShell-Yaml is available
                         '\.[json|jsonc]'
                         {
                             $jsonFile = Get-Content -Raw -Path $configFile
@@ -337,7 +371,7 @@ process
     }
 }
 
-Begin
+begin
 {
     # Find build config if not specified.
     if (-not $BuildConfig)
@@ -451,7 +485,8 @@ Begin
 
     if ($ResolveDependency)
     {
-        Write-Host -Object "[pre-build] Resolving dependencies." -ForegroundColor Green
+        Write-Host -Object "[pre-build] Resolving dependencies using preferred method." -ForegroundColor Green
+
         $resolveDependencyParams = @{ }
 
         # If BuildConfig is a Yaml file, bootstrap powershell-yaml via ResolveDependency.
@@ -501,13 +536,6 @@ Begin
             Write-Verbose -Message "Dependency already resolved. Removing task."
 
             $null = $PSBoundParameters.Remove('ResolveDependency')
-        }
-
-        if ($Env:ModuleVersion) {
-            Write-Host "Module version is already set to '$Env:ModuleVersion'" -ForegroundColor 'Yellow'
-        } elseif (Get-Command 'gitversion' -ErrorAction 'SilentlyContinue') {
-            $Env:ModuleVersion = (gitversion /output json /format "{SemVer}") -replace '^([\d\.]+\-\w+)\.(\d+)$', '$1-$2'
-            Write-Host "Setting module version to the output of 'gitversion': '$Env:ModuleVersion'." -ForegroundColor 'Yellow'
         }
 
         Write-Host -Object "[build] Starting build with InvokeBuild." -ForegroundColor Green
